% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roi2polygon.R
\name{roi2polygon}
\alias{roi2polygon}
\title{roi2polygon: Transform a '.roi' files created with ImageJ to a spatial 
polygons.}
\usage{
roi2polygon(roi.folder, pic.folder)
}
\arguments{
\item{roi.folder}{path to the folder that contains the '.roi' files. roi 
files could be located in subfolders.}

\item{pic.folder}{path to the folder that contains the image files. 
The function will take the first image of the folder as reference. 
All images in the pic.folder must have the same dimensions.}
}
\value{
A list of spatial polygons
}
\description{
Read '.roi' files located in a folder and subfolders and transform to a list 
of spatial polygons.
}
\examples{

roi2polygon.2 <- function(roi.folder = "./myrois", pic.folder = "./mypics")

}
\author{
Manuel Molina-Bustamante
}
