% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart2.R
\name{chart.2}
\alias{chart.2}
\title{Manual selection of the image regions where we can find the color checker tiles.
You need to click over color tiles in the opened window where you can find the picture with the color chart. 
To click over the tiles you need to folow the order shown in
\href{https://raw.githubusercontent.com/MMolBus/photomoss/master/vignettes/vignette_Photomoss_workflow/chart.png}{this image}.}
\usage{
chart.2(pic.path, samp.width = 0.01, pic.format)
}
\arguments{
\item{samp.width}{numeric. Distance from original click point to establish the perimeter of the new geometry. Of length 1 replicated to the number of input click points, or of length equal to the number of click points.}

\item{pic.format}{character. Picture file format. It could be "jpg" for .jpg,.JPG and .jpeg; or "tif", for .tif format.}

\item{x}{string. File path where you can find the image files.}
}
\value{
A raster with 24 features one by each color tile.
}
\description{
Manual selection of the image regions where we can find the color checker tiles.
You need to click over color tiles in the opened window where you can find the picture with the color chart. 
To click over the tiles you need to folow the order shown in
\href{https://raw.githubusercontent.com/MMolBus/photomoss/master/vignettes/vignette_Photomoss_workflow/chart.png}{this image}.
}
\examples{

chart.2(pic.path="./JPG", samp.width = 0.01, pic.format = "jpg")

}
\author{
Manuel Molina-Bustamante
}
